Move fold implementation to separate module
